pyproject.toml
[tool.poetry]
authors = ["Hju Kneyck Flores <hjucode@gmail.com>"]
description = ""
name = "app"
version = "0.1.0"

[tool.poetry.dependencies]
PyNaCl = "^1.5.0"
aiohttp = "^3.8.1"
base58 = "^2.1.1"
blacksheep = "^1.2.7"
cryptography = "^37.0.4"
orjson = "^3.7.12"
pydantic = {extras = ["email", "dotenv"], version = "^1.9.2"}
python = "^3.10"
solana = "^0.25.1"
uvicorn = "^0.18.2"
uvloop = "^0.16.0"
xxhash = "^3.0.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
pylint = "^2.14.5"
pytest = "^5.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
